third_party/blink/web_tests/mojo/lite/constants.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="file:///gen/mojo/public/js/ts/bindings/tests/constants.test-mojom-lite.js"></script>
<script>
'use strict';

[{
  constant: mojo.tstest.kBoolValue,
  expectedValue: true,
  type: 'boolean'
}, {
  constant: mojo.tstest.kInt8Value,
  expectedValue: -2,
  type: 'int8'
}, {
  constant: mojo.tstest.kUint8Value,
  expectedValue: 128,
  type: 'uint8'
}, {
  constant: mojo.tstest.kInt16Value,
  expectedValue: -233,
  type: 'int16'
}, {
  constant: mojo.tstest.kUint16Value,
  expectedValue: 44204,
  type: 'uint16'
}, {
  constant: mojo.tstest.kInt32Value,
  expectedValue: -44204,
  type: 'int32'
}, {
  constant: mojo.tstest.kUint32Value,
  expectedValue: 4294967295,
  type: 'uint32'
}, {
  // TODO(crbug.com/1009262): Re-enable once we generate the right value for
  // 64-bit constants.
  //
  //  constant: mojo.tstest.kInt64Value,
  //  expectedValue: -9223372036854775807n,
  //  type: 'int64'
  //}, {
  //  constant: mojo.tstest.kUint64Value,
  //  expectedValue: 9999999999999999999n,
  //  type: 'uint64'
  //}, {
  constant: mojo.tstest.kDoubleValue,
  expectedValue: 3.14159,
  type: 'double'
}, {
  constant: mojo.tstest.kDoubleInfinity,
  expectedValue: Infinity,
  type: 'double infinity'
}, {
  constant: mojo.tstest.kDoubleNegativeInfinity,
  expectedValue: -Infinity,
  type: 'double negative infinity'
}, {
  constant: mojo.tstest.kDoubleNaN,
  expectedValue: NaN,
  type: 'double NaN'
}, {
  constant: mojo.tstest.kFloatValue,
  expectedValue: 2.71828,
  type: 'float'
}, {
  constant: mojo.tstest.kFloatInfinity,
  expectedValue: Infinity,
  type: 'float infinity'
}, {
  constant: mojo.tstest.kFloatNegativeInfinity,
  expectedValue: -Infinity,
  type: 'float negative infinity'
}, {
  constant: mojo.tstest.kFloatNaN,
  expectedValue: NaN,
  type: 'NaN'
}, {
  constant: mojo.tstest.kStringValue,
  expectedValue: "test string contents",
  type: 'string'
}].forEach(testCase => {
  test(
    () => assert_equals(testCase.constant, testCase.expectedValue),
    `Checks that the value of generated ${testCase.type} constants is correct`);
});
</script>
